improved the pathing and added visualization of individuums on the grid
Pathing works so far, right now everything is mostly manually tested. Test cases need to be improved a little more?

connected game logic with the game engine. Now we can visualize Individuums on the grid and start get things moving (meaning that literally)

Next steps should include a class to handle rounds and the game loop and controls with mouse and stuff